calendar_integration/requirements.txt
google-auth==1.8.0
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.3
google-api-python-client==1.7.2
django-cors-headers
django-sslserver
djangorestframework